+++ src/nvupdater/update_manager_dialog.py
@@ -25,25 +25,28 @@ class UpdateManagerDialog(ModalDialog, UpdateManagerCtrl):
 
         self.title(f'{_("Check for updates")} - novelibre @release')
 
+        self.protocol("WM_DELETE_WINDOW", self.on_quit)
+
         columns = 'Module', 'Installed version', 'Latest version'
-        self.moduleCollection = ttk.Treeview(self, columns=columns, show='headings', selectmode='browse')
+        self.repoList = ttk.Treeview(self, columns=columns, show='headings', selectmode='browse')
 
         # scrollY = ttk.Scrollbar(self.moduleCollection, orient='vertical', command=self.moduleCollection.yview)
         # self.moduleCollection.configure(yscrollcommand=scrollY.set)
         # scrollY.pack(side='right', fill='y')
         #--- unsolved problem: adding a scollbar makes the window shrink to minimum
 
-        self.moduleCollection.pack(fill='both', expand=True)
-        self.moduleCollection.bind('<<TreeviewSelect>>', self.on_select_module)
-        self.moduleCollection.tag_configure('rejected', foreground='red')
-        self.moduleCollection.tag_configure('inactive', foreground='gray')
+        self.repoList.pack(fill='both', expand=True)
+        self.repoList.bind('<<TreeviewSelect>>', self.on_select_module)
+        self.repoList.tag_configure('outdated', foreground='red')
+        self.repoList.tag_configure('updated', foreground='green')
+        self.repoList.tag_configure('inactive', foreground='gray')
 
-        self.moduleCollection.column('Module', width=150, minwidth=120, stretch=False)
-        self.moduleCollection.heading('Module', text=_('Module'), anchor='w')
-        self.moduleCollection.column('Installed version', width=100, minwidth=100, stretch=False)
-        self.moduleCollection.heading('Installed version', text=_('Installed version'), anchor='w')
-        self.moduleCollection.column('Latest version', width=100, minwidth=100, stretch=False)
-        self.moduleCollection.heading('Latest version', text=_('Latest version'), anchor='w')
+        self.repoList.column('Module', width=150, minwidth=120, stretch=False)
+        self.repoList.heading('Module', text=_('Module'), anchor='w')
+        self.repoList.column('Installed version', width=100, minwidth=100, stretch=False)
+        self.repoList.heading('Installed version', text=_('Installed version'), anchor='w')
+        self.repoList.column('Latest version', width=100, minwidth=100, stretch=False)
+        self.repoList.heading('Latest version', text=_('Latest version'), anchor='w')
 
         self.messagingArea = tk.Label(self, fg='white', bg='green')
         self.messagingArea.pack(fill='x')
@@ -73,7 +76,7 @@ class UpdateManagerDialog(ModalDialog, UpdateManagerCtrl):
         ttk.Button(
             self._footer,
             text=_('Close'),
-            command=self.destroy
+            command=self.on_quit
             ).pack(padx=5, pady=5, side='right')
 
         # "Help" button.
@@ -85,10 +88,14 @@ class UpdateManagerDialog(ModalDialog, UpdateManagerCtrl):
 
         # Set Key bindings.
         self.bind(KEYS.OPEN_HELP[0], NvupdaterHelp.open_help_page)
+        self.repoList.bind('<Double-1>', self.update_module)
 
         # Populate the list.
         self.build_module_list()
 
     def output(self, text):
-        self.messagingArea.configure(text=text)
+        try:
+            self.messagingArea.configure(text=text)
+        except:
+            pass