+++ docs/changelog.md
@@ -4,6 +4,15 @@
 
 ## Changelog
 
+
+### Version 5.1.0
+
+- Showing a modal dialog with a list like the plugin manager (#1).
+
+API: 5.0
+Based on novelibre 5.16.3
+
+
 ### Version 5.0.4
 
 - Refactored the code for better maintainability.